WIP: improve lifecycle
